+++ src/filter_cache.h
@@ -33,9 +33,6 @@
 #include "ooo_core_recorder.h"
 #include "lbr.h"
 
-#include <list>
-#include <unordered_map>
-
 /* Extends Cache with an L0 direct-mapped cache, optimized to hell for hits
  *
  * L1 lookups are dominated by several kinds of overhead (grab the cache locks,
@@ -45,75 +42,6 @@
  * and then go through the normal access path. Because there is one line per set,
  * it is fine to do this without grabbing a lock.
  */
-
-struct prefetched_info
-{
-    Address vAddrLine;
-    uint64_t availCycle;
-    prefetched_info(Address _vAddrLine, uint64_t _availCycle)
-    {
-        vAddrLine = _vAddrLine;
-        availCycle = _availCycle;
-    }
-};
-
-
-class PrefetchBuffer
-{
-private:
-    uint64_t capacity;
-    std::list<prefetched_info> dq;
-    std::unordered_map<Address,std::list<prefetched_info>::iterator> index;
-public:
-    PrefetchBuffer(uint64_t size)
-    {
-        capacity = size;
-    }
-    bool prefetch(Address addr, uint64_t availCycle)
-    {
-        bool was_present;
-        uint64_t effectiveAvailCycle = availCycle;
-        if(index.find(addr)==index.end())
-        {
-            was_present = false;
-            if(likely(dq.size()==capacity))
-            {
-                Address to_be_removed = dq.back().vAddrLine;
-                dq.pop_back();
-                index.erase(to_be_removed);
-            }
-        }
-        else
-        {
-            was_present = true;
-            effectiveAvailCycle = MIN(index[addr]->availCycle, effectiveAvailCycle);
-            dq.erase(index[addr]);
-        }
-        dq.push_front(prefetched_info(addr, effectiveAvailCycle));
-        index[addr]=dq.begin();
-        return was_present;
-    }
-    bool transfer(Address addr, uint64_t &availCycle)
-    {
-        if(index.find(addr)==index.end())
-        {
-            return false;
-        }
-        availCycle = index[addr]->availCycle;
-        dq.erase(index[addr]);
-        index.erase(addr);
-        return true;
-    }
-    void clear()
-    {
-        index.clear();
-        dq.clear();
-    }
-    ~PrefetchBuffer()
-    {
-        clear();
-    }
-};
 
 class FilterCache : public Cache {
     protected:
@@ -146,7 +74,6 @@ class FilterCache : public Cache {
                 : addr(_addr), skip(_skip), pc(_pc), isSW(_isSW), serialize(_serialize) {};
         };
         g_vector<PrefetchInfo> prefetchQueue;
-        PrefetchBuffer *prefetch_buffer = nullptr;
 
     public:
         FilterCache(uint32_t _numSets, uint32_t _numLines, CC* _cc, CacheArray* _array,
@@ -211,25 +138,7 @@ class FilterCache : public Cache {
                 fGETSHit++;
                 return MAX(curCycle, availCycle);
             } else {
-                if((prefetch_buffer!=nullptr) && prefetch_buffer->transfer(vLineAddr, availCycle))
-                {
-                    replace(vLineAddr, idx, true, curCycle, pc, lbr, no_update_timestamp);
-                    return MAX(curCycle+accLat,availCycle);
-                }
-                else
-                {
-                    return replace(vLineAddr, idx, true, curCycle, pc, lbr, no_update_timestamp);
-                }
-            }
-        }
-
-        inline void prefetch_into_buffer(Address vAddr, uint64_t curCycle)
-        {
-            const uint64_t prefetch_cost = 27;
-            Address vLineAddr = vAddr >> lineBits;
-            if(prefetch_buffer!=nullptr)
-            {
-                prefetch_buffer->prefetch(vLineAddr, curCycle+prefetch_cost);
+                return replace(vLineAddr, idx, true, curCycle, pc, lbr, no_update_timestamp);
             }
         }
 
@@ -243,15 +152,7 @@ class FilterCache : public Cache {
                 //filterArray[idx].availCycle = curCycle; //do optimistic store-load forwarding
                 return MAX(curCycle, availCycle);
             } else {
-                if((prefetch_buffer!=nullptr) && prefetch_buffer->transfer(vLineAddr, availCycle))
-                {
-                    replace(vLineAddr, idx, false, curCycle, pc);
-                    return MAX(curCycle+accLat, availCycle);
-                }
-                else
-                {
-                    return replace(vLineAddr, idx, false, curCycle, pc);
-                }
+                return replace(vLineAddr, idx, false, curCycle, pc);
             }
         }
 
@@ -340,22 +241,6 @@ class FilterCache : public Cache {
             futex_lock(&filterLock);
             for (uint32_t i = 0; i < numSets; i++) filterArray[i].clear();
             futex_unlock(&filterLock);
-        }
-
-        void setPrefetchBuffer(uint64_t capacity)
-        {
-            if(prefetch_buffer==nullptr)
-            {
-                prefetch_buffer = new PrefetchBuffer(capacity);
-            }
-        }
-        void clearPrefetchBuffer()
-        {
-            if(prefetch_buffer!=nullptr)
-            {
-                prefetch_buffer->clear();
-                delete prefetch_buffer;
-            }
         }
 
     private:
